contrib/cloud.google.com/go/pubsub.v1: move tracing related code to a separate package
