+++ contrib/cloud.google.com/go/pubsub.v1/option.go
@@ -6,41 +6,17 @@
 package pubsub
 
 import (
-	"gopkg.in/DataDog/dd-trace-go.v1/internal/namingschema"
+	"gopkg.in/DataDog/dd-trace-go.v1/contrib/cloud.google.com/go/pubsub.v1/internal/tracing"
 )
 
-type config struct {
-	serviceName     string
-	publishSpanName string
-	receiveSpanName string
-	measured        bool
-}
+// Option is used to customize spans started by WrapReceiveHandler or Publish.
+type Option = tracing.Option
 
-func defaultConfig() *config {
-	return &config{
-		serviceName:     namingschema.ServiceNameOverrideV0("", ""),
-		publishSpanName: namingschema.OpName(namingschema.GCPPubSubOutbound),
-		receiveSpanName: namingschema.OpName(namingschema.GCPPubSubInbound),
-		measured:        false,
-	}
-}
-
-// A Option is used to customize spans started by WrapReceiveHandler or Publish.
-type Option func(cfg *config)
-
-// A ReceiveOption has been deprecated in favor of Option.
+// Deprecated: ReceiveOption has been deprecated in favor of Option.
 type ReceiveOption = Option
 
 // WithServiceName sets the service name tag for traces started by WrapReceiveHandler or Publish.
-func WithServiceName(serviceName string) Option {
-	return func(cfg *config) {
-		cfg.serviceName = serviceName
-	}
-}
+var WithServiceName = tracing.WithServiceName
 
 // WithMeasured sets the measured tag for traces started by WrapReceiveHandler or Publish.
-func WithMeasured() Option {
-	return func(cfg *config) {
-		cfg.measured = true
-	}
-}
+var WithMeasured = tracing.WithMeasured
